Move the decorators module into core.
